test: 인수 테스트 픽스쳐 추상 클래스로 추출하여 중복 제
