+++ src/test/java/roomescape/acceptance/AcceptanceFixture.java
@@ -1,0 +1,41 @@
+package roomescape.acceptance;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.test.context.ActiveProfiles;
+
+import io.restassured.RestAssured;
+import roomescape.service.security.JwtProvider;
+
+@ActiveProfiles("test")
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+public abstract class AcceptanceFixture {
+    @LocalServerPort
+    private int port;
+    @Autowired
+    protected JdbcTemplate jdbcTemplate;
+    @Autowired
+    protected JwtProvider jwtProvider;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+    }
+
+    @AfterEach
+    void tearDown() {
+        jdbcTemplate.update("DELETE FROM reservation");
+        jdbcTemplate.update("ALTER TABLE reservation ALTER COLUMN id RESTART");
+        jdbcTemplate.update("DELETE FROM reservation_time");
+        jdbcTemplate.update("ALTER TABLE reservation_time ALTER COLUMN id RESTART");
+        jdbcTemplate.update("DELETE FROM theme");
+        jdbcTemplate.update("ALTER TABLE theme ALTER COLUMN id RESTART");
+        jdbcTemplate.update("DELETE FROM member");
+        jdbcTemplate.update("ALTER TABLE member ALTER COLUMN id RESTART");
+    }
+}
