+++ src/test/java/roomescape/acceptance/ReservationAcceptanceTest.java
@@ -6,50 +6,16 @@ import java.time.LocalDate;
 import java.util.Map;
 
 import org.apache.http.HttpStatus;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.context.ActiveProfiles;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import roomescape.domain.Member;
 import roomescape.domain.Role;
-import roomescape.service.security.JwtProvider;
 
-@ActiveProfiles("test")
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-class ReservationAcceptanceTest {
-    @LocalServerPort
-    private int port;
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-    @Autowired
-    private JwtProvider jwtProvider;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
-
-    @AfterEach
-    void tearDown() {
-        jdbcTemplate.update("DELETE FROM reservation");
-        jdbcTemplate.update("ALTER TABLE reservation ALTER COLUMN id RESTART");
-        jdbcTemplate.update("DELETE FROM reservation_time");
-        jdbcTemplate.update("ALTER TABLE reservation_time ALTER COLUMN id RESTART");
-        jdbcTemplate.update("DELETE FROM theme");
-        jdbcTemplate.update("ALTER TABLE theme ALTER COLUMN id RESTART");
-        jdbcTemplate.update("DELETE FROM member");
-        jdbcTemplate.update("ALTER TABLE member ALTER COLUMN id RESTART");
-    }
-
+class ReservationAcceptanceTest extends AcceptanceFixture {
+    
     @Test
     @DisplayName("예약 생성 API")
     void member_generate_API() {
